requirements.txt
flask==2.3.2
sentence-transformers==2.2.2
torch>=1.12.0
nltk==3.8.1
python-docx==0.8.11
PyPDF2==3.0.1
